requirements.txt
cloudscraper>=1.2.71
tqdm>=4.66.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
fastmcp>=2.0.0
PyYAML>=6.0
